Start on types assembly.
Changing the model to simplify usage by having only one namespace and class structure for both WinRT and desktop. This will allow moving functionality into the WinRT version without name chages as the Windows store whitelists additional APIs.

The WInterop.Types assembly will include all of the data types and the bare minimum of P/Invokes that are needed for things such as closing handles. This will allow exchanging of objects between assemblies that are built against either the desktop or the WinRT version of the WInterop library.