First major step to consuming TerraFX interop
There isn't as much value in maintaining P/Invoke definitions with auto-generated libraries. This starts moving everything towards
@TannerGooding's TerraFX windows interop library.

It currently consumes daily builds as it depends on recently
added wrappers. I'll move it to nuget feeds once it is possible.

Goal is to remove all P/Invokes and COM interface definitions from
this library. Manually defining the P/Invokes was tedious and somewhat risky. COM interop in .NET is slow and cannot actually
correctly map things (notably with DirectX).

This will hopefully improve my velocity with this library so there will
be more examples to code to. As specified in the readme, I hope to continue to expand on my "WinForms" lite idea.